+++ billing-service/src/main/java/com/billing/service/dto/response/StockUpdateDTO.java
@@ -2,8 +2,10 @@ package com.billing.service.dto.response;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class StockUpdateDTO {
     private Long stockId;
-    private Integer newQty;
+    private BigDecimal newQty;
 }
