+++ returns-service/src/main/java/com/returns/service/dto/request/BillingItemRequestDTO.java
@@ -3,11 +3,12 @@ package com.returns.service.dto.request;
 import lombok.Data;
 
 import java.math.BigDecimal;
+import java.math.BigDecimal;
 
 
 @Data
 public class BillingItemRequestDTO {
-    private int qty;
+    private BigDecimal qty;
     private BigDecimal salesPrice;
     private Double salesDiscount;
     private Long stock;
