+++ returns-service/src/main/java/com/returns/service/dto/request/CustomerReturnRequestItemDTO.java
@@ -3,9 +3,11 @@ package com.returns.service.dto.request;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
+import java.math.BigDecimal;
+
 @EqualsAndHashCode(callSuper = true)
 @Data
 public class CustomerReturnRequestItemDTO extends ChannelRequestDTO{
     private String itemCode;
-    private int qty;
+    private BigDecimal qty;
 }
